+++ tests/ccd/test_ccd.cpp
@@ -1,6 +1,9 @@
 #include <catch2/catch.hpp>
 
+#include <ipc/ipc.hpp>
 #include <ipc/ccd/ccd.hpp>
+
+#include <test_utils.hpp>
 
 TEST_CASE("Vertex-Vertex Impact", "[ccd]")
 {
@@ -18,4 +21,38 @@ TEST_CASE("Vertex-Vertex Impact", "[ccd]")
 
     REQUIRE(is_collision);
     CHECK(toi == Approx(0.5));
+}
+
+TEST_CASE("Repeated CCD", "[ccd][thisone]")
+{
+    Eigen::MatrixXd V0, V1;
+    Eigen::MatrixXi E, F;
+
+    bool success = load_mesh("ccd-failure/0.obj", V0, E, F);
+    if (!success) {
+        return;
+    }
+    // REQUIRE(success);
+
+    success = load_mesh("ccd-failure/1.obj", V1, E, F);
+    if (!success) {
+        return;
+    }
+    // REQUIRE(success);
+
+    bool has_collisions = !ipc::is_step_collision_free(V0, V1, E, F);
+
+    double stepsize = ipc::compute_collision_free_stepsize(V0, V1, E, F);
+
+    fmt::print(
+        "has_collisions={} stepsize={:.17g}\n", has_collisions, stepsize);
+
+    bool has_collisions2 =
+        !ipc::is_step_collision_free(V0, (V1 - V0) * stepsize + V0, E, F);
+
+    double stepsize2 = ipc::compute_collision_free_stepsize(
+        V0, (V1 - V0) * stepsize + V0, E, F);
+
+    fmt::print(
+        "has_collisions2={} stepsize2={:.17g}\n", has_collisions2, stepsize2);
 }